refactor: Moving SideMenu to components
